moved constants to common.py and models to model.py
