+++ servicenow/stackstate_checks/servicenow/__init__.py
@@ -2,12 +2,13 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from .__about__ import __version__
-from .servicenow import ServicenowCheck, InstanceInfo, State, json_parse_exception
+from .model import State, InstanceInfo
+from .servicenow import ServicenowCheck, json_parse_exception
 
 __all__ = [
     '__version__',
     'ServicenowCheck',
-    'InstanceInfo',
+    'json_parse_exception',
     'State',
-    'json_parse_exception'
+    'InstanceInfo'
 ]
